requirements.txt
alive-progress~=1.5.1
certifi~=2020.4.5.2
chardet~=3.0.4
colorama~=0.4.3
idna~=2.9
Pillow~=7.2.0
psutil~=5.7.0
PyPDF2~=1.26.0
PyPDF4~=1.27.0
qrcode~=6.1
reportlab~=3.5.42
requests~=2.24.0
six~=1.15.0
urllib3~=1.25.9
progress~=1.5
python-dotenv~=0.14.0